requirements.txt
alembic==1.17.2
bcrypt==5.0.0
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.7
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
Platypus==1.0
ply==3.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.4
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
xlsxwriter==3.2.9
